:sparkles: delete reservation api
