+++ src/main/java/msg/team1/Hi/domain/homebase/presentation/controller/admin/AdminHomeBaseController.java
@@ -1,14 +1,12 @@
-package msg.team1.Hi.domain.home_base.presentation.controller.student;
+package msg.team1.Hi.domain.homebase.presentation.controller.admin;
 
 import lombok.RequiredArgsConstructor;
-import msg.team1.Hi.domain.home_base.presentation.dto.request.ReserveHomeBaseRequest;
-import msg.team1.Hi.domain.home_base.presentation.dto.request.UpdateReservationMemberRequest;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationDetailResponse;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
-import msg.team1.Hi.domain.home_base.service.ReserveHomeBaseService;
+import msg.team1.Hi.domain.homebase.presentation.dto.request.ReserveHomeBaseRequest;
+import msg.team1.Hi.domain.homebase.presentation.dto.response.LookUpReservationDetailResponse;
+import msg.team1.Hi.domain.homebase.presentation.dto.response.LookUpReservationResponse;
+import msg.team1.Hi.domain.homebase.service.ReserveHomeBaseService;
 import msg.team1.Hi.domain.reservation.service.GetAllReservationService;
 import msg.team1.Hi.domain.reservation.service.GetReservationService;
-import msg.team1.Hi.domain.reservation.service.UpdateReservationMemberService;
 import msg.team1.Hi.domain.reservation.service.UpdateReservationTeamNameService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,14 +17,13 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/student/home-base")
-public class StudentHomeBaseController {
+@RequestMapping("/admin/home-base")
+public class AdminHomeBaseController {
 
     private final ReserveHomeBaseService reserveHomeBaseService;
     private final GetAllReservationService getAllReservationService;
     private final GetReservationService getReservationService;
     private final UpdateReservationTeamNameService updateReservationTeamNameService;
-    private final UpdateReservationMemberService updateReservationMemberService;
 
     @PostMapping("/reserve")
     public ResponseEntity<Void> reserveHomeBase(@Valid @RequestBody ReserveHomeBaseRequest request) {
@@ -48,14 +45,8 @@ public class StudentHomeBaseController {
     }
 
     @PatchMapping("/{reservation_id}/name")
-    public ResponseEntity<Void> updateReservationTeamName(@PathVariable("reservation_id") Long reservationId, @RequestParam String teamName){
+    public ResponseEntity<Void> updateReservationTeamName(@PathVariable("reservation_id") Long reservationId, @RequestParam String teamName) {
         updateReservationTeamNameService.execute(reservationId, teamName);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PatchMapping("/{reservation_id}/update-member")
-    public ResponseEntity<Void> updateReservationMembers(@PathVariable("reservation_id") Long reservationId, @RequestBody @Valid UpdateReservationMemberRequest request) {
-        updateReservationMemberService.execute(request, reservationId);
         return ResponseEntity.noContent().build();
     }
 }
