+++ src/main/java/msg/team1/Hi/domain/homebase/presentation/controller/student/StudentHomeBaseController.java
@@ -1,14 +1,12 @@
-package msg.team1.Hi.domain.home_base.presentation.controller.admin;
+package msg.team1.Hi.domain.homebase.presentation.controller.student;
 
 import lombok.RequiredArgsConstructor;
-import msg.team1.Hi.domain.home_base.presentation.dto.request.ReserveHomeBaseRequest;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationDetailResponse;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
-import msg.team1.Hi.domain.home_base.service.ReserveHomeBaseService;
-import msg.team1.Hi.domain.reservation.service.GetAllReservationService;
-import msg.team1.Hi.domain.reservation.service.GetReservationService;
-import msg.team1.Hi.domain.reservation.service.UpdateReservationMemberService;
-import msg.team1.Hi.domain.reservation.service.UpdateReservationTeamNameService;
+import msg.team1.Hi.domain.homebase.presentation.dto.request.ReserveHomeBaseRequest;
+import msg.team1.Hi.domain.homebase.presentation.dto.request.UpdateReservationMemberRequest;
+import msg.team1.Hi.domain.homebase.presentation.dto.response.LookUpReservationDetailResponse;
+import msg.team1.Hi.domain.homebase.presentation.dto.response.LookUpReservationResponse;
+import msg.team1.Hi.domain.homebase.service.ReserveHomeBaseService;
+import msg.team1.Hi.domain.reservation.service.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,13 +16,15 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/admin/home-base")
-public class AdminHomeBaseController {
+@RequestMapping("/student/home-base")
+public class StudentHomeBaseController {
 
     private final ReserveHomeBaseService reserveHomeBaseService;
     private final GetAllReservationService getAllReservationService;
     private final GetReservationService getReservationService;
     private final UpdateReservationTeamNameService updateReservationTeamNameService;
+    private final UpdateReservationMemberService updateReservationMemberService;
+    private final DeleteReservationService deleteReservationService;
 
     @PostMapping("/reserve")
     public ResponseEntity<Void> reserveHomeBase(@Valid @RequestBody ReserveHomeBaseRequest request) {
@@ -46,8 +46,20 @@ public class AdminHomeBaseController {
     }
 
     @PatchMapping("/{reservation_id}/name")
-    public ResponseEntity<Void> updateReservationTeamName(@PathVariable("reservation_id") Long reservationId, @RequestParam String teamName) {
+    public ResponseEntity<Void> updateReservationTeamName(@PathVariable("reservation_id") Long reservationId, @RequestParam String teamName){
         updateReservationTeamNameService.execute(reservationId, teamName);
         return ResponseEntity.noContent().build();
+    }
+
+    @PatchMapping("/{reservation_id}/update-member")
+    public ResponseEntity<Void> updateReservationMembers(@PathVariable("reservation_id") Long reservationId, @RequestBody @Valid UpdateReservationMemberRequest request) {
+        updateReservationMemberService.execute(request, reservationId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping("/{reservation_id}/delete")
+    public ResponseEntity<Void> deleteReservation(@PathVariable("reservation_id") Long reservationId){
+        deleteReservationService.execute(reservationId);
+        return ResponseEntity.ok().build();
     }
 }
