+++ src/main/java/msg/team1/Hi/domain/homebase/repository/HomeBaseRepository.java
@@ -1,6 +1,6 @@
-package msg.team1.Hi.domain.home_base.repository;
+package msg.team1.Hi.domain.homebase.repository;
 
-import msg.team1.Hi.domain.home_base.entity.HomeBase;
+import msg.team1.Hi.domain.homebase.entity.HomeBase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
