+++ src/main/java/msg/team1/Hi/domain/reservation/entity/Reservation.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import msg.team1.Hi.domain.home_base.entity.HomeBase;
+import msg.team1.Hi.domain.homebase.entity.HomeBase;
 import msg.team1.Hi.domain.member.entity.Member;
 import msg.team1.Hi.domain.reservation.entity.enum_type.CheckStatus;
