+++ src/main/java/msg/team1/Hi/global/error/handler/GlobalExceptionHandler.java
@@ -5,10 +5,10 @@ import msg.team1.Hi.domain.email.exception.AuthCodeExpiredException;
 import msg.team1.Hi.domain.email.exception.ManyRequestEmailAuthException;
 import msg.team1.Hi.domain.email.exception.MisMatchAuthCodeException;
 import msg.team1.Hi.domain.email.exception.NotVerifyEmailException;
-import msg.team1.Hi.domain.home_base.exception.AlreadyReservedException;
+import msg.team1.Hi.domain.homebase.exception.AlreadyReservedException;
 import msg.team1.Hi.domain.auth.exception.ExistEmailException;
 import msg.team1.Hi.domain.auth.exception.MemberNotFoundException;
-import msg.team1.Hi.domain.home_base.exception.FullHomeBaseReservationException;
+import msg.team1.Hi.domain.homebase.exception.FullHomeBaseReservationException;
 import msg.team1.Hi.domain.member.exception.MisMatchPasswordException;
 import msg.team1.Hi.domain.auth.exception.RefreshTokenNotFoundException;
 import msg.team1.Hi.domain.notice.exception.NoticeNotFoundException;
