+++ src/main/java/msg/team1/Hi/global/util/HomeBaseUtil.java
@@ -1,8 +1,8 @@
 package msg.team1.Hi.global.util;
 
 import lombok.RequiredArgsConstructor;
-import msg.team1.Hi.domain.home_base.presentation.dto.response.LookUpReservationResponse;
-import msg.team1.Hi.domain.home_base.repository.HomeBaseRepository;
+import msg.team1.Hi.domain.homebase.presentation.dto.response.LookUpReservationResponse;
+import msg.team1.Hi.domain.homebase.repository.HomeBaseRepository;
 import msg.team1.Hi.domain.reservation.entity.Reservation;
 import org.springframework.stereotype.Component;
